simplified code to rtk standard
